+++ recipes/auto_chef.rb
@@ -4,7 +4,7 @@
 #
 # Author:: Vassilis Rizopoulos (<var@zuehlke.com>)
 #
-# Copyright (c) 2014-2016 Zühlke, All Rights Reserved.
+# Copyright (c) 2014-2018 Zühlke, All Rights Reserved.
 
 #Create a user to restart the process after booting